MLOPS PROJECT - 6/PROJECT - 6/requirements.txt
numpy
pandas
setuptools
mysql-connector-python
scikit-learn

